+++ CHANGELOG.md
@@ -53,6 +53,8 @@ Coq 8.13, 8.14, and 8.15 are no longer supported.
 * Add monotone resource algebra, `algebra/mra.v`, to enable reasoning about
   monotonicity with respect to an arbitrary preorder relation: the extension order
   of `mra R` is designed to embed the preorder relation `R`. (by Amin Timany)
+* Rename instances `union_with_proper` → `union_with_ne`,
+  `map_fmap_proper` → `map_fmap_ne`, `map_zip_with_proper` → `map_zip_with_ne`.
 
 **Changes in `bi`:**